refactor(followerstatus)
